Prepare mask for inclusion of multiple vendors
* Use shifts for mask and make it into ssize_t (so it gets arithmetic
  shift instead of logical)
* Remove hardcoded names of nvidia in variable names outside of nvidia
  file.

+++ include/nvtop/extract_gpuinfo.h
@@ -26,7 +26,7 @@
 
 #include <stdbool.h>
 
-bool gpuinfo_init_info_extraction(uint64_t mask, unsigned *devices_count, struct list_head *devices);
+bool gpuinfo_init_info_extraction(ssize_t mask, unsigned *devices_count, struct list_head *devices);
 
 bool gpuinfo_shutdown_info_extraction(struct list_head *devices);
